requirements.txt
python-telegram-bot==20.4
rich
colorama
apscheduler
openai
sqlite3